requirements.txt
Flask
groq
python-dotenv
pillow
pytesseract
